requirements.txt
twikit 
python-dotenv 
rich 
fastapi
uvicorn[standard] 
requests 